backend/requirements.txt
python-dotenv
requests
pydantic
bs4
pymongo[srv]
fastapi
uvicorn